Made the details of assignment parts customizable
You can now provide any number of descriptions for assignment parts in
the parts_conf.txt file. The program reads them at startup, creating a
list you can choose from. Each assignment part can have a name, a filter
for the filenames of submissions, and a filter for the filename of the
official grade file.